Fix a bug in GeneralConfigWidget that was reverting the grid optins everytime the user applyed settings
Crow's foot implementation checkpoint

+++ libobjrenderer/src/baseobjectview.cpp
@@ -631,14 +631,14 @@ void BaseObjectView::togglePlaceholder(bool visible)
 	}
 }
 
-float BaseObjectView::getFontFactor(void)
+double BaseObjectView::getFontFactor(void)
 {
 	return(font_config[ParsersAttributes::GLOBAL].font().pointSizeF()/DEFAULT_FONT_SIZE);
 }
 
-float BaseObjectView::getScreenDpiFactor(void)
+double BaseObjectView::getScreenDpiFactor(void)
 {
-	float factor = qApp->screens().at(qApp->desktop()->screenNumber(qApp->activeWindow()))->logicalDotsPerInch() / 96.0f;
+	double factor = qApp->screens().at(qApp->desktop()->screenNumber(qApp->activeWindow()))->logicalDotsPerInch() / 96.0f;
 
 	if(factor < 1)
 		return (1);
